+++ proxy/proxy.go
@@ -15,14 +15,30 @@ func CreateReverseProxy(targetStr string, timeout time.Duration, keepAlive time.
 	if err != nil {
 		return nil, fmt.Errorf("cannot parse target, %s", err)
 	}
+	// director := func(req *http.Request) {
+	// 	req.URL = target
+	// 	for key, value := range headers {
+	// 		req.Header[key] = value
+	// 	}
+
+	// 	// TODO: set x forwarded headers
+	// }
+
 	director := func(req *http.Request) {
+		// force custom endpoint
 		req.URL = target
+
+		req.Host = target.Host
+		if ip, _, err := net.SplitHostPort(req.RemoteAddr); err == nil {
+			req.Header.Set("X-Forwarded-For", ip)
+		}
+		req.Header.Set("X-Forwarded-Proto", req.URL.Scheme)
+		req.Header.Set("X-Forwarded-Host", req.Host)
 		for key, value := range headers {
 			req.Header[key] = value
 		}
-
-		// TODO: set x forwarded headers
 	}
+
 	return &httputil.ReverseProxy{Director: director, Transport: &http.Transport{
 		Proxy: http.ProxyFromEnvironment,
 		Dial: (&net.Dialer{