+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All changes will be reported in this file
 
+## [v0.0.53] - 2025-12-14
+
+### Fix
+
+- Fixed reverse proxy config
+
+## [v0.0.52] - 2025-12-14
+
+### Changed
+
+- Changed reverse proxy config
+
 ## [v0.0.51] - 2025-11-28
 
 ### Added